Update window_textbox.js
First pass at trying to support automatic "sanitization" of modified functions that patches may load. The method in use, however, is flawed: it compares function text from a single patch source file against the current function text of the function in the loaded instance, which may be an amalgamation of several plugins. Thus, if a core/non-plugin script (eg, rpg_core.js) is modified, nothing will appear in the js, and it will consider a function that isn't actually modified in rpg_core, that should "match" its original function definition and thus be pruned, won't match because in the loaded instance a plugin has overwritten that function. But since the patch's version of rpg_core reports a "change" in that function, it will overwrite the overwrite, resetting the function that a plugin modified to match what was originally in rpg_core.