Update mv_mm.js
Added first attempt at "replacing instances of globally-scoped let/const declarations with var declarations" to avoid crashing when duplicate plugins are loaded. Creates a one-time popup menu.
I'm worried that this will: 1) be intimidating or confusing, 2) be inaccurate, as people creating mods may perform this check, add their own "unclean" scripts, and never run it again. Which will likely not even introduce an issue as such plugins would be new to the patch, and not be loaded multiple times, but still.
I will transition it to fit the silent, automated nature of the rest of the design paradigm, simply performing the check whenever a patch is created or a an install is exported. I considered keeping the receipt file and using its contents as a list of "cleaned" files to save runtime, but these exports are expected to be length, and people could modify a js file after it is "cleaned".