Update mv_mm.js
Fixed an issue with the var-replacement regarding the ordering, to make it consistent between diffing and patching (so that no index mismatches occur between diffs and patch)
specifically, patch was applying diffs and then doing the replacement on the project afterwards, which means the first set of diff applications was done with a diff made from a cleaned version of the game, onto a dirty version, and that all subsequent applications/patches would be cleaned diffs onto cleaned base game files. Changed it to clean the base game repository the first time any "apply patch" is called, *before* it tries to apply any diffs.